Create a non-root user in container to use for ssh.
If the image is run using the run.sh script then the user is given
the same user name, uid, group name and gid as the current host user.
In this way any files owned by either user on a shared volume will
be owned by both.

If the run command is _not_ given any environment variables to specify
the user data then it defaults to user and group names of dev and
uid:gid 1000:1000.